implemented offshore beacon
the offshore beacon is able to setup the mesh network by determining their timings for them and sending that out in a singal.
after setting them up the offshore beacon can request for information when it needs to then the mesh will send information back as to what it has collected. then the mesh will clear its data it has been holding.