Setup most of the new protocol for the beacon communication
The beacons now connect to each other and have actual connections within the mesh.
when sending information through the mesh they send more direct messages towards certain beacons to traverse through the mesh and make it to the master.

beacons can disconnect and reconnect to and from the mesh automatically